Fixed issue where line number metadata would be freed as type uint8_t instead of int.

+++ src/compiler/chunk.c
@@ -55,8 +55,8 @@ void writeChunk(Chunk* chunk, uint8_t byte, int line){
 
 void freeChunk(Chunk* chunk){
 	FREE_ARRAY(uint8_t, chunk->code, chunk->capacity);
-	FREE_ARRAY(uint8_t, chunk->opsPerLine, chunk->lineCapacity);
-	FREE_ARRAY(uint8_t, chunk->lineNums, chunk->lineCapacity);
+	FREE_ARRAY(int, chunk->opsPerLine, chunk->lineCapacity);
+	FREE_ARRAY(int, chunk->lineNums, chunk->lineCapacity);
 	freeValueArray(&chunk->constants);
 }
